# openapi.yaml
openapi: 3.0.0
info:
    title: RE Remake Interactive Map API
    description: |
        RESTful API for querying Resident Evil Remake game data including items, biohazards, maps, areas, and rooms.

        All endpoints support filtering by various parameters and include rate limiting (100 requests per 15 minutes).

        ## Features
        - Query resources by IDs with batch support
        - Search and filter by multiple criteria
        - S.T.A.R.S. biohazard rankings
        - Comprehensive room and map data

        ## Rate Limiting
        All endpoints are rate-limited to 100 requests per 15-minute window. Rate limit information is provided in response headers.
    version: 1.0.0
    contact:
        name: RE Remake Interactive Map
        url: https://github.com/UncaughtTypeError/RE-Remake-Interactive-Map
servers:
    - url: http://localhost:3000
      description: Local development server
    - url: https://your-production-url.com
      description: Production server
...
paths:
...
    /api/biohazards/all:
        get:
            tags: [Biohazards]
            summary: Get all biohazards (base data)
            description: |
                Retrieves all biohazards (enemies) base data without any filters. Returns the complete biohazard database with master biohazard definitions including names, types, and S.T.A.R.S. codes.
            responses:
                '200':
                    description: Successful response with all base biohazards
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/BiohazardData'
                '429':
                    $ref: '#/components/responses/RateLimitError'
    /api/biohazards:
        get:
            tags: [Biohazards]
            summary: Get biohazards by IDs (base data)
            description: |
                Retrieves base biohazard definitions by their unique BiohazardCode IDs. Supports comma-separated batch queries (max 100 IDs).
                Returns both found biohazards and any unrecognized IDs. This returns master biohazard definitions, not room-specific instances.
            parameters:
                - name: ids
                  in: query
                  required: false
                  schema:
                      type: string
                      example: 'Zb,Ht,Cb'
                  description: Comma-separated list of BiohazardCode IDs (e.g., 'Zb' for Zombie, 'Ht' for Hunter, 'Cb' for Cerberus, 'Ad' for Adder). Max 100 IDs.
            responses:
                '200':
                    description: Successful response with found base biohazards and unrecognized IDs
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    foundBiohazards:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/BiohazardData'
                                    unrecognizedIds:
                                        type: array
                                        items:
                                            type: string
                '400':
                    $ref: '#/components/responses/ValidationError'
                '404':
                    $ref: '#/components/responses/NotFoundError'
                '429':
                    $ref: '#/components/responses/RateLimitError'
...
    /api/biohazards/search:
        get:
            tags: [Biohazards]
            summary: Search and filter biohazards (base data)
            description: |
                Search base biohazards by S.T.A.R.S. code or name. Returns master biohazard definitions without room locations or complete S.T.A.R.S. intelligence.
                All parameters are optional and can be combined. Returns biohazards matching ALL specified criteria (AND logic).
            parameters:
                - name: code
                  in: query
                  schema:
                      type: string
                      example: 'Zb'
                  description: Filter by S.T.A.R.S. code (e.g., 'Zb' for Zombie)
                - name: name
                  in: query
                  schema:
                      type: string
                      example: 'Zombie'
                  description: Filter by biohazard name (partial match, case-insensitive)
            responses:
                '200':
                    description: Successful response with filtered base biohazards
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/BiohazardData'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '429':
                    $ref: '#/components/responses/RateLimitError'
    /api/biohazards/rooms/all:
        get:
            tags: [Biohazards]
            summary: Get all biohazards (room data)
            description: |
                Retrieves all biohazards with complete room location data and S.T.A.R.S. (Statistical Threat Analysis Rating Standard) intelligence.
                Includes threat classifications (low-severe), Greek rankings (Eta-Alpha), star ratings (1-3), engagement directives,
                authorized response procedures, effective measures (weapon damage/shot counts), threat profiles, behavioral patterns,
                and caution advisories for field operatives.
            responses:
                '200':
                    description: Successful response with all room biohazards
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/BiohazardRoomData'
                '429':
                    $ref: '#/components/responses/RateLimitError'
...
    /api/biohazards/rooms/search:
        get:
            tags: [Biohazards]
            summary: Search and filter biohazards (room data)
            description: |
                Search biohazards with complete room location data by room, difficulty level, S.T.A.R.S. code, or name. All parameters are optional.
                Returns biohazards matching ALL specified criteria (AND logic). Includes complete S.T.A.R.S. (Statistical Threat Analysis Rating Standard) intelligence:
                threat classifications (low-severe), Greek rankings (Eta-Alpha), star ratings (1-3), engagement directives, authorized response procedures,
                effective measures (weapon damage/shot counts), threat profiles, behavioral patterns, and caution advisories for field operatives.
            parameters:
                - name: room
                  in: query
                  schema:
                      type: string
                      example: 'keepersRoom'
                  description: Filter by room ID (e.g., 'keepersRoom', 'teaRoomCorridor', 'mainHallF1')
                - name: difficulty
                  in: query
                  schema:
                      type: string
                      enum:
                          [
                              JV-lvl-very-easy,
                              JV-lvl-easy,
                              JV-lvl-normal,
                              JV-lvl-hard,
                              CR-lvl-very-easy,
                              CR-lvl-easy,
                              CR-lvl-normal,
                              CR-lvl-hard,
                          ]
                      example: 'JV-lvl-normal'
                  description: Filter by difficulty level
                - name: code
                  in: query
                  schema:
                      type: string
                      example: 'Zb'
                  description: Filter by S.T.A.R.S. BiohazardCode (e.g., 'Zb' for Zombie, 'Ht' for Hunter, 'Cb' for Cerberus, 'Ad' for Adder)
                - name: name
                  in: query
                  schema:
                      type: string
                      example: 'Zombie'
                  description: Filter by biohazard name (partial match, case-insensitive)
            responses:
                '200':
                    description: Successful response with filtered room biohazards
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/BiohazardRoomData'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '429':
                    $ref: '#/components/responses/RateLimitError'

    # ============================================================================
    # MAPS ENDPOINTS - AREAS
    # ============================================================================
...
components:
    schemas:
...
        BiohazardData:
            type: object
            description: Base biohazard data (master biohazard definition)
            properties:
                id:
                    type: string
                    description: BiohazardCode - unique 2-letter identifier
                    example: 'Zb'
                name:
                    type: string
                    description: Biohazard name
                    example: 'Zombie'
                codeName:
                    type: string
                    nullable: true
                    description: Code name if applicable
                    example: null
                threatLevel:
                    type: string
                    description: Threat level assessment
                    example: 'Low-Moderate'
                starsClassification:
                    type: string
                    description: S.T.A.R.S. classification (Greek letter name)
                    example: 'Zeta'
                starsRanking:
                    type: string
                    description: S.T.A.R.S. ranking number
                    example: '0Half'
                directive:
                    type: string
                    description: Tactical directive
                    example: 'Avoid else Engage'
                prevalence:
                    type: string
                    description: How common this biohazard is
                    example: 'Common'
                imageSrc:
                    type: string
                    description: URL to biohazard image
                    example: 'https://s-media-cache-ak0.pinimg.com/originals/f1/26/57/f126578050920fc271fd8b33534a9605.jpg'
                isVenomous:
                    type: boolean
                    description: Whether the biohazard is venomous
                    example: false
                authorizedResponseProcedures:
                    type: array
                    items:
                        type: string
                    description: Recommended response procedures
                    example: ['Engage from secure distances', 'Avoid close proximity']
                effectiveMeasures:
                    type: string
                    description: Effective weapons and measures
                    example: 'short to long-range small or medium arms, defensive weapons, fire'
                caution:
                    type: string
                    nullable: true
                    description: Caution advisory
                    example: 'Low-Moderate Risk threat level in close proximity with numbers in excess of 2.'
                threats:
                    type: array
                    items:
                        type: object
                        properties:
                            threat:
                                type: string
                            description:
                                type: string
                    description: Specific threats posed
                    example:
                        [
                            {
                                threat: 'Mutation Threat',
                                description: 'Crimson Head mutation, V-ACT (Virus-Activation) secondary infection stage',
                            },
                        ]
                characteristics:
                    type: array
                    items:
                        type: string
                    description: Behavioral characteristics
                    example: ['slow-moving', 'sedate', 'unperceptive']
...
        BiohazardRoomData:
            type: object
            description: Biohazard (enemy) data with room location information
            properties:
                id:
                    type: string
                    description: Unique biohazard instance identifier (format - biohazardType-roomId)
                    example: 'zombie1-keepersRoom'
                name:
                    type: string
                    description: Biohazard name
                    example: 'Zombie'
                code:
                    type: string
                    description: S.T.A.R.S. BiohazardCode - references BiohazardData.id
                    example: 'Zb'
                qty:
                    type: integer
                    description: Number of enemies
                    example: 1
                ambush:
                    type: boolean
                    nullable: true
                    description: Whether this is an ambush encounter
                    example: false
                isFeatured:
                    type: boolean
                    description: Whether this biohazard is featured
                    example: false
                map:
                    type: object
                    properties:
                        roomId:
                            type: string
                            example: 'keepersRoom'
                        mapId:
                            type: string
                            example: 'mansionF1'
                difficultyLevel:
                    type: array
                    items:
                        type: string
                        enum:
                            [
                                JV-lvl-very-easy,
                                JV-lvl-easy,
                                JV-lvl-normal,
                                JV-lvl-hard,
                                CR-lvl-very-easy,
                                CR-lvl-easy,
                                CR-lvl-normal,
                                CR-lvl-hard,
                            ]
                    description: Difficulty levels where this biohazard appears
                    example:
                        [
                            'JV-lvl-very-easy',
                            'JV-lvl-easy',
                            'JV-lvl-normal',
                            'JV-lvl-hard',
                            'CR-lvl-very-easy',
                            'CR-lvl-easy',
                            'CR-lvl-normal',
                            'CR-lvl-hard',
                        ]
...
        Error:
            type: object
            properties:
                error:
                    type: string
                    description: Error message
                    example: 'Invalid query parameters'
                details:
                    type: array
                    items:
                        type: string
                    description: Detailed error information
                    example: ['ids must be comma-separated alphanumeric strings']
    responses:
        ValidationError:
            description: Invalid request parameters
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
                    example:
                        error: 'Invalid query parameters'
                        details: ['ids must be comma-separated alphanumeric strings']
        NotFoundError:
            description: Resource not found
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
                    example:
                        error: 'No IDs provided are recognized'
        RateLimitError:
            description: Rate limit exceeded
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            message:
                                type: string
                                example: 'Too many requests, please try again later.'